+++ modules/render/src/texture_data.cpp
@@ -24,7 +24,7 @@
 namespace argus {
 
     // IMPORTANT: image_data is assumed to be allocated on the heap
-    TextureData::TextureData(const size_t width, const size_t height, unsigned char **&&image_data):
+    TextureData::TextureData(const unsigned int width, const unsigned int height, unsigned char **&&image_data):
             pimpl(new pimpl_TextureData()),
             width(width),
             height(height) {
